create reusable LabelText and Legend form components
