+++ components/src/Form/LabelText.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+const LabelText = styled.span(({ theme }) => ({
+  fontSize: theme.fontSizes.small,
+  fontWeight: theme.fontWeights.bold,
+  lineHeight: theme.lineHeights.smallTextBase
+}));
+
+export default LabelText;
