+++ components/src/Form/Legend.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+const Legend = styled.legend(({ theme }) => ({
+  marginBottom: theme.space.x1
+}));
+
+export default Legend;
