+++ components/src/Form/index.js
@@ -2,3 +2,5 @@ export { default as Field } from "./Field";
 export { default as Fieldset } from "./Fieldset";
 export { default as Form } from "./Form";
 export { default as FormSection } from "./FormSection";
+export { default as LabelText } from "./LabelText";
+export { default as Legend } from "./Legend";
