+++ components/src/Radio/RadioGroup.js
@@ -1,17 +1,11 @@
-import React, { useContext } from "react";
+import React from "react";
 import PropTypes from "prop-types";
-import styled, { ThemeContext } from "styled-components";
+import styled from "styled-components";
 
 import Radio from "./Radio";
 import { HelpText, RequirementText } from "../FieldLabel";
 import { InlineValidation } from "../Validation";
-import { Fieldset } from "../Form";
-
-const labelTextStyles = theme => ({
-  fontSize: theme.fontSizes.small,
-  fontWeight: theme.fontWeights.bold,
-  lineHeight: theme.lineHeights.smallTextBase
-});
+import { Fieldset, Legend, LabelText } from "../Form";
 
 const getRadioButtons = props => {
   const radioButtons = React.Children.map(props.children, radio => {
@@ -35,14 +29,12 @@ const getRadioButtons = props => {
 
 const BaseRadioGroup = ({ className, id, errorMessage, errorList, labelText, helpText, requirementText, ...props }) => {
   const otherProps = { ...props, errorMessage, errorList };
-  const themeContext = useContext(ThemeContext);
-
   return (
     <Fieldset className={className} id={id} hasHelpText={!!helpText}>
-      <legend style={{ marginBottom: themeContext.space.x1 }}>
-        <span style={labelTextStyles(themeContext)}>{labelText}</span>
+      <Legend>
+        <LabelText>{labelText}</LabelText>
         {requirementText && <RequirementText>{requirementText}</RequirementText>}
-      </legend>
+      </Legend>
       {helpText && <HelpText>{helpText}</HelpText>}
       {getRadioButtons(otherProps)}
       <InlineValidation mt="x1" errorMessage={errorMessage} errorList={errorList} />
