+++ components/src/Toggle/Toggle.js
@@ -1,25 +1,18 @@
-import React, { useState, useContext } from "react";
+import React, { useState } from "react";
 import PropTypes from "prop-types";
-import styled, { ThemeContext } from "styled-components";
+import styled from "styled-components";
 import { Box } from "../Box";
 import { HelpText, RequirementText } from "../FieldLabel";
-import { Field } from "../Form";
+import { Field, LabelText } from "../Form";
 import { Text } from "../Type";
 import { ClickInputLabel } from "../utils";
 import ToggleButton from "./ToggleButton";
 
-const labelTextStyles = theme => ({
-  fontSize: theme.fontSizes.small,
-  fontWeight: theme.fontWeights.bold,
-  lineHeight: theme.lineHeights.smallTextBase
-});
-
 const MaybeToggleTitle = ({ labelText, requirementText, helpText, children, ...props }) => {
-  const themeContext = useContext(ThemeContext);
   return labelText ? (
     <div {...props}>
       <Box mb={children && "x1"}>
-        <span style={labelTextStyles(themeContext)}>{labelText}</span>
+        <LabelText>{labelText}</LabelText>
         {requirementText && <RequirementText>{requirementText}</RequirementText>}
         {helpText && <HelpText>{helpText}</HelpText>}
       </Box>
